Tools/_CP14/ColorReplacer/requirements.txt
pillow~=11.0.0